# By u214892 (48) and others
# Via AutoMerge User for PullRequest (10) and others
* commit '6cba275a52460bba355c526ff479eec3c4de4e50': (91 commits)
  Added libXext
  WZU-2994: build optimizer enabled
  WZU-2994: better implementation of build optimizer
  Modified exportpath
  WZU-3017: disabled build optimizer because feature branch was not built
  Added IIB9 Binaries
  WZU-2994: Build optimizer: enabled
  WZU-2994: Build optimizer: array is not empty when no file edited.
  WZU-2994: first implementation of a build optimizer.
  WZU-2994: fixed bad pattern in registry select case
  WZU-2994 Docker-Pipeline über interne Docker-Registry
  WZU-2994: export confluence log directory in init script
  WZU-2994: fix for not valid pull request id
  WZU-2994 Docker-Pipeline über interne Docker-Registry
  WZU-2994 Docker-Pipeline über interne Docker-Registry
  WZU-2994 Docker-Pipeline über interne Docker-Registry
  WZU-2994: rework of registry shell scripts
  WZU-2994: added missing sql file for external stash
  WZU-2994: +x for remove-container.sh
  WZU-2994: image deployment improved
  ...

+++ _scripts/remove-container.sh
@@ -12,5 +12,7 @@ fsid=`sudo docker inspect -f='{{.Name}}' ${containerid} | awk -F"-" '{print $NF 
 sudo docker rm $containerid
 
 # remove fs
-sudo rm -rf $externalfshome/$fsid
+if [[ ! -z "$fsid" ]]; then
+    sudo rm -rf $externalfshome/$fsid
+fi
